# openapi/contacts.json
{
  "components": {
    "schemas": {
...
      "ActivityRead": {
        "properties": {
          "activity_type": {
            "default": "note",
            "maxLength": 40,
            "minLength": 1,
            "title": "Activity Type",
            "type": "string"
          },
          "body": {
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "contact_id": {
            "title": "Contact Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "created_at",
          "updated_at",
          "body",
          "contact_id",
          "created_by"
        ],
        "title": "ActivityRead",
        "type": "object"
      },
      "ActivityUpdate": {
        "properties": {
          "activity_type": {
            "anyOf": [
              {
                "maxLength": 40,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activity Type"
          },
          "body": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          }
        },
        "title": "ActivityUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Contacts, addresses, communication channels, tags, activity, and memberships.",
    "title": "OpenVend Contacts Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/activity/{activity_id}": {
      "delete": {
        "operationId": "delete_activity_activity__activity_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "activity_id",
            "required": true,
            "schema": {
              "title": "Activity Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-OpenVend-Scopes",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Openvend-Scopes"
            }
          },
          {
            "in": "header",
            "name": "X-OpenVend-Actor-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Openvend-Actor-Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Activity",
        "tags": [
          "activity"
        ]
      },
      "patch": {
        "operationId": "update_activity_activity__activity_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "activity_id",
            "required": true,
            "schema": {
              "title": "Activity Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-OpenVend-Scopes",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Openvend-Scopes"
            }
          },
          {
            "in": "header",
            "name": "X-OpenVend-Actor-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Openvend-Actor-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivityUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Activity",
        "tags": [
          "activity"
        ]
      }
    },
    "/addresses/{address_id}": {
      "delete": {
        "operationId": "delete_address_addresses__address_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "address_id",
            "required": true,
            "schema": {
              "title": "Address Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-OpenVend-Scopes",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Openvend-Scopes"
            }
          },
          {
            "in": "header",
            "name": "X-OpenVend-Actor-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Openvend-Actor-Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Address",
        "tags": [
          "addresses"
        ]
      },
      "patch": {
        "operationId": "update_address_addresses__address_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "address_id",
            "required": true,
            "schema": {
              "title": "Address Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-OpenVend-Scopes",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Openvend-Scopes"
            }
          },
          {
            "in": "header",
            "name": "X-OpenVend-Actor-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Openvend-Actor-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Address",
        "tags": [
          "addresses"
        ]
      }
    },
    "/communication-channels/{channel_id}": {
      "delete": {
        "operationId": "delete_communication_channel_communication_channels__channel_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "channel_id",
            "required": true,
            "schema": {
              "title": "Channel Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-OpenVend-Scopes",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Openvend-Scopes"
            }
          },
          {
            "in": "header",
            "name": "X-OpenVend-Actor-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Openvend-Actor-Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Communication Channel",
        "tags": [
          "communication-channels"
        ]
      },
      "patch": {
        "operationId": "update_communication_channel_communication_channels__channel_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "channel_id",
            "required": true,
            "schema": {
              "title": "Channel Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-OpenVend-Scopes",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Openvend-Scopes"
            }
          },
          {
            "in": "header",
            "name": "X-OpenVend-Actor-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Openvend-Actor-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommunicationChannelUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunicationChannelRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Communication Channel",
        "tags": [
          "communication-channels"
        ]
      }
    },
...
  }
}